refactor: idx구별에 따른 테스트코드 수정
